requirements.txt
django>=6.0,<6.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
gunicorn>=23.0,<24.0
psycopg[binary]>=3.2,<3.3
whitenoise>=6.8,<6.9
